Add user_helpers module for test helpers
